Moved Builder mode toggle to admin panel
The Builder mode was previously controlled via the configuration file. This was done to make it harder for people to break their production instances.

Having it in the configuration file is cumbersome however and required a server restart to apply. We also couldn´t easily add warnings inside the configuration file regarding the Builder mode.

Moving it into the admin panel makes it easier to enable, but also gives us the ability to warn the admin user. It can also show more clearly the dependency between maintenance mode and the Builder (one requires the other).