Added preset record ID as query filter
Before we had to filter to (hopefully) fixed preset names; with this change, we can filter to presets directly.

To safely use preset filters, they should only ever refer to protected presets; otherwise we filter to potentially non-existing record IDs. The Builder only shows protected presets in corresponding inputs, but presets can be set to 'unprotected' afterwards. There is a fallback in place but the filter will not work as intended if the referenced preset record does not exist anymore.