Added default list limit as valid limit option
By default there is a number of list limit options (10, 25, 50, etc.). The default limit from the list field does not have to conform with these options however.

When a list default limit was set that was not a valid option, the limit input would appear empty. With this change the default list limit is added to the list of allowed limit options, if it is not already on it.