Improved form record handling with missing access
If form attempts to load a record but gets no return, previously we just forwarded to an empty form. This is not desirable if we work on a record, save and then loose access directly because of the changed values (due to relation or form filters).

Now when a record is loaded (or reloaded after save) the form will keep its state but switch all input fields to readonly. This means that when saving a record and loosing access because of it, the record values stay visible with the whole form switching to readonly. The user sees the updated values but cannot continue working on the record.