Changed NULL value gallery column handling in calendar fields
Before, if a files attribute that serves as a gallery column had a NULL value, a placeholder image would be shown. This worked but was inconsistent with how gallery columns worked in list views; in lists NULL value galleries would not be shown at all, as it could lead to large lists showing placeholders for all records.

The optimal solution would be to offer configuration options, whether to use placeholder images (and which one). Until then, this changes the behavior to be in line with list views. Calendar fields (and Gantts correspondingly) will now ignore columns with NULL value galleries.