Fixed default field value being overwritten
If the default value for an input field was an array, its object reference was assigned to the field value map instead of its value. Because of this it was overwritten with the current field value and was not available when a new record was being created. By cloning the object, this issue is avoided.