Yet another cache denial edge case
The PWA reliably reloads its cache when it starts. But if the PWA is running, the server disconnects, updates and then reconnects, the PWA does not initiate a reload.

This commit forces the PWA to execute a reload if the returned application version from the server on new connection is different than a previously known, locally stored application version.