Fixed: File cleanup task removes wrong files
Wrong files were identified to be cleaned up in some cases - this is now fixed. However the cleanup job can currently not identify all required files and will therefore leave some out. A new solution is required.

As deletion of wrong files is critical, we will release this version. The fix for the cleanup job is not critical and can occur later.