Reduced unneeded trigger options
Reduced trigger options constraint, deferrable & initially deferred down to 'deferred', which now toggles all three settings simultaneously.

As DB transactions are handled by the system and the app author has no control over them, the only real difference is whether a trigger is executed deferred or not; this only occurs if all 3 options are active and is ignored if a single one is disabled. Therefore the other options are not needed and were removed.