Fixed: Invalid index during change log creation
Because file attribute values are handled differently to regular values, checking for changes for the data log needed to work around this. This was quite complex and had a bug.

The issue occurred because the number of old values being checked was different to the number of all attributes (because file attribute value retrieval is unnecessary for change detection as file attribute values already include changes).

The fix: File attributes are still not retrieved, but to simplify the comparison, a NULL value is used as placeholder; this way the total value count is always correct.