UI blocker for schema update on client
For very slow connections, between requesting and applying a new schema, the client UI should be blocked.

The issue never came up because the process is usually instant. On connections with a very low bandwidth with large schema it might take longer than a couple of seconds however; without block, a client might attempt changes while a new, incompatible schema is applied afterwards.