get_record_id(): Support bulk forms
Allow frontend function "get_record_id()" to retrieve open record IDs for the current form, if used in bulk editing mode.

Because bulk editing forms do not retrieve record values, record IDs of joined relations are not available. In most cases, we need the base relation so this is still useful. Updated form context help to reflect this restriction.