Dynamic. reduce header elements for avail. space
Before, we had a three-version-header: Regular, reduced and mobile.

Now the header dynamically adjusts elements following available space and login settings. The more space is available, the more elements can be shown fully - with less space, elements are reduced or replaced with smaller alternatives.

This solution is more flexible and required for planned login options and features.